backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
opencv-python-headless>=4.8.0
imagehash>=4.3.1
numpy>=1.24.0
python-multipart>=0.0.6
mnemonic>=0.20
pynacl>=1.5.0
pillow>=10.0.0
